requirements.txt
Django==2.0.3
dj-database-url==0.4.2
django-allauth==0.35.0
django-appconf==1.0.2
django-bootstrap4==0.0.6
django-concurrency==1.4
django-debug-toolbar==1.9.1
django-timezone-field==2.1
environs==2.1.0
psycopg2==2.7.4
pytz==2018.3
waitress==1.1.0
whitenoise==3.3.1
